Guess grid with locations of data-points in NREL's collection

Guess grid with locations of data-points in NREL's collection

Usage

guess_nrel_grid(
  gis_sf,
  dx = 0.01,
  dy = 0.01,
  q_atributes = "windspeed_100m",
  collection = "wtk",
  q_interval = 60,
  q_names = "2014",
  steps_max = 10,
  verbose = T
)

Arguments

gis_sf

sf object (map) of the region to estimate a number of data-points in NREL's collection, and their coordinates (grid).

dx

starting step along longitude

dy

starting step along latitude

q_atributes

attributes of the query (see fetch_nrel_data)

collection

name of NREL's data-collection (see get_nrel_url)

q_interval

interval parameter in the query

q_names

names parameter in the query

steps_max

maximum steps in evaluation of the grid along each axis, 10 by default

verbose

logical, should the process be reported

Value

list with the grid in sf points format, grid parameters, fetched initial data-points used for the estimation of the grid.